requirements.txt
Flask==2.3.3
requests==2.31.0
gunicorn==21.2.0
google-api-python-client==2.108.0
isodate==0.6.1
google-ads==22.1.0
pandas==2.0.3
python-dotenv==1.0.0